improve test parallelization
